memory improvements

+++ src/Player.cpp
@@ -1,6 +1,7 @@
 #include "../headers/Player.hpp"
 #include <SFML/Window/Keyboard.hpp>
 #include <string>
+#include <sys/types.h>
 #include "../headers/path.hpp"
 
 Player::Player()
@@ -15,25 +16,25 @@ Player::Player()
     imgL.loadFromFile(filepath+"sprites/norbi/idle/norbiL.png");
 
     // walking
-    for (int i = 1; i <= 2; i++)
+    for (u_int8_t i = 1; i <= 2; i++)
     {
         imgUW[i - 1].loadFromFile(filepath+"sprites/norbi/walk/norbiU" + std::to_string(i) + ".png");
     }
-    for (int i = 1; i <= 2; i++)
+    for (u_int8_t i = 1; i <= 2; i++)
     {
         imgDW[i - 1].loadFromFile(filepath+"sprites/norbi/walk/norbiD" + std::to_string(i) + ".png");
     }
-    for (int i = 1; i <= 2; i++)
+    for (u_int8_t i = 1; i <= 2; i++)
     {
         imgRW[i - 1].loadFromFile(filepath+"sprites/norbi/walk/norbiR" + std::to_string(i) + ".png");
     }
-    for (int i = 1; i <= 2; i++)
+    for (u_int8_t i = 1; i <= 2; i++)
     {
         imgLW[i - 1].loadFromFile(filepath+"sprites/norbi/walk/norbiL" + std::to_string(i) + ".png");
     }
 
     //sounds and sound buffers
-    for(int i=0; i<2; i++){
+    for(u_int8_t i=0; i<2; i++){
         orzelB[i].loadFromFile(filepath+"audio/orzel/orzel"+std::to_string(i+1)+".wav");
         orzel[i].setBuffer(orzelB[i]);
     }
